Add JWT token validation to CourseController

+++ src/main/java/dev/hiruna/edutrack/controller/CourseController.java
@@ -2,7 +2,10 @@ package dev.hiruna.edutrack.controller;
 
 import dev.hiruna.edutrack.dto.CourseDTO;
 import dev.hiruna.edutrack.service.CourseService;
+import dev.hiruna.edutrack.util.JWTAuthenticator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,28 +17,50 @@ public class CourseController {
     @Autowired
     private CourseService courseService;
 
+    @Autowired
+    private JWTAuthenticator jwtAuthenticator;
+
     @GetMapping
-    public List<CourseDTO> getAllCourses() {
-        return courseService.getAllCourses();
+    public ResponseEntity<List<CourseDTO>> getAllCourses(@RequestHeader("Authorization") String authHeader) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            return new ResponseEntity<>(courseService.getAllCourses(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/{id}")
-    public CourseDTO getCourseById(@PathVariable Integer id) {
-        return courseService.getCourseById(id);
+    public ResponseEntity<CourseDTO> getCourseById(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            CourseDTO courseDTO = courseService.getCourseById(id);
+            return new ResponseEntity<>(courseDTO, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PostMapping
-    public CourseDTO createCourse(@RequestBody CourseDTO courseDTO) {
-        return courseService.createCourse(courseDTO);
+    public ResponseEntity<CourseDTO> createCourse(@RequestHeader("Authorization") String authHeader, @RequestBody CourseDTO courseDTO) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            CourseDTO createdCourse = courseService.createCourse(courseDTO);
+            return new ResponseEntity<>(createdCourse, HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping("/{id}")
-    public CourseDTO updateCourse(@PathVariable Integer id, @RequestBody CourseDTO courseDTO) {
-        return courseService.updateCourse(id, courseDTO);
+    public ResponseEntity<CourseDTO> updateCourse(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id, @RequestBody CourseDTO courseDTO) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            CourseDTO updatedCourse = courseService.updateCourse(id, courseDTO);
+            return new ResponseEntity<>(updatedCourse, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCourse(@PathVariable Integer id) {
-        courseService.deleteCourse(id);
+    public ResponseEntity<Void> deleteCourse(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            courseService.deleteCourse(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
